@@ -1,6 +1,6 @@
 module github.com/kopia/kopia/site
 
-go 1.20
+go 1.21
 
 require (
 	github.com/google/docsy v0.7.0 // indirect
