@@ -11,9 +11,10 @@ import (
 )
 
 type commandACLAdd struct {
-	user   string
-	target string
-	level  string
+	user      string
+	target    string
+	level     string
+	overwrite bool
 }
 
 func (c *commandACLAdd) setup(svc appServices, parent commandParent) {
@@ -21,6 +22,7 @@ func (c *commandACLAdd) setup(svc appServices, parent commandParent) {
 	cmd.Flag("user", "User the ACL targets").Required().StringVar(&c.user)
 	cmd.Flag("target", "Manifests targeted by the rule (type:T,key1:value1,...,keyN:valueN)").Required().StringVar(&c.target)
 	cmd.Flag("access", "Access the user gets to subject").Required().EnumVar(&c.level, acl.SupportedAccessLevels()...)
+	cmd.Flag("overwrite", "Overwrite existing rule with the same user and target").BoolVar(&c.overwrite)
 	cmd.Action(svc.repositoryWriterAction(c.run))
 }
 
@@ -47,5 +49,5 @@ func (c *commandACLAdd) run(ctx context.Context, rep repo.RepositoryWriter) erro
 		Access: al,
 	}
 
-	return errors.Wrap(acl.AddACL(ctx, rep, e), "error adding ACL entry")
+	return errors.Wrap(acl.AddACL(ctx, rep, e, c.overwrite), "error adding ACL entry")
 }
