@@ -30,13 +30,21 @@ func TestCacheSet(t *testing.T) {
 		"cache", "set",
 		"--cache-directory", ncd,
 		"--content-cache-size-mb=33",
+		"--content-cache-size-limit-mb=331",
 		"--metadata-cache-size-mb=44",
+		"--metadata-cache-size-limit-mb=441",
 	)
 
 	out := env.RunAndExpectSuccess(t, "cache", "info")
 	require.Contains(t, mustGetLineContaining(t, out, "33 MB"), ncd)
-	require.Contains(t, mustGetLineContaining(t, out, "33 MB"), "contents")
-	require.Contains(t, mustGetLineContaining(t, out, "44 MB"), "metadata")
+	require.Contains(t, mustGetLineContaining(t, out, "soft limit: 33 MB"), "contents")
+	require.Contains(t, mustGetLineContaining(t, out, "hard limit: 331 MB"), "contents")
+	require.Contains(t, mustGetLineContaining(t, out, "min sweep age: 10m0s"), "contents")
+
+	require.Contains(t, mustGetLineContaining(t, out, "soft limit: 44 MB"), "metadata")
+	require.Contains(t, mustGetLineContaining(t, out, "hard limit: 441 MB"), "metadata")
+	require.Contains(t, mustGetLineContaining(t, out, "min sweep age: 24h0m0s"), "metadata")
+
 	require.Contains(t, mustGetLineContaining(t, out, "55s"), "blob-list")
 }
 
