@@ -14,12 +14,13 @@ import (
 	"github.com/kopia/kopia/repo/blob"
 )
 
-var ErrBlobLocked = errors.New("cannot alter object before retention period expires")
+var ErrBlobLocked = errors.New("cannot alter object before retention period expires") // +checklocksignore
 
 type entry struct {
 	value          []byte
 	mtime          time.Time
 	retentionTime  time.Time
+	retentionMode  blob.RetentionMode
 	isDeleteMarker bool
 }
 
@@ -31,6 +32,7 @@ type versionedEntries map[blob.ID][]*entry
 // marker. This struct manages the retention time of each blob through the
 // PutBlob options.
 type objectLockingMap struct {
+	blob.DefaultProviderImplementation
 	// +checklocks:mutex
 	data    versionedEntries
 	timeNow func() time.Time // +checklocksignore
@@ -66,10 +68,6 @@ func (s *objectLockingMap) getLatestForMutationLocked(id blob.ID) (*entry, error
 	}
 
 	return e, nil
-}
-
-func (s *objectLockingMap) GetCapacity(ctx context.Context) (blob.Capacity, error) {
-	return blob.Capacity{}, blob.ErrNotAVolume
 }
 
 // GetBlob works the same as map-storage GetBlob except that if the latest
@@ -129,6 +127,18 @@ func (s *objectLockingMap) GetMetadata(ctx context.Context, id blob.ID) (blob.Me
 	}, nil
 }
 
+func (s *objectLockingMap) GetRetention(ctx context.Context, id blob.ID) (blob.RetentionMode, time.Time, error) {
+	s.mutex.RLock()
+	defer s.mutex.RUnlock()
+
+	e, err := s.getLatestByID(id)
+	if err != nil {
+		return "", time.Time{}, errors.Wrap(err, "getting blob")
+	}
+
+	return e.retentionMode, e.retentionTime, nil
+}
+
 // PutBlob works the same as map-storage PutBlob except that if the latest
 // version is a delete-marker then it will return ErrBlobNotFound. The
 // PutOptions retention parameters will be respected when storing the object.
@@ -153,6 +163,7 @@ func (s *objectLockingMap) PutBlob(ctx context.Context, id blob.ID, data blob.By
 
 	if opts.HasRetentionOptions() {
 		e.retentionTime = e.mtime.Add(opts.RetentionPeriod)
+		e.retentionMode = opts.RetentionMode
 	}
 
 	s.data[id] = append(s.data[id], e)
@@ -199,8 +210,13 @@ func (s *objectLockingMap) ExtendBlobRetention(ctx context.Context, id blob.ID, 
 		return blob.ErrBlobNotFound
 	}
 
+	// Update the retention time from now to the given retention period in the
+	// future. Note that we do not bump the existing time on the element `e.mtime`
+	// by the given delta because we'd like to align with the S3 storage's current
+	// time and the S3 storage code extends retention periods based off the
+	// current time, not object mod time.
 	if !e.retentionTime.IsZero() {
-		e.retentionTime = e.mtime.Add(opts.RetentionPeriod)
+		e.retentionTime = s.timeNow().Add(opts.RetentionPeriod)
 	}
 
 	return nil
@@ -243,11 +259,6 @@ func (s *objectLockingMap) ListBlobs(ctx context.Context, prefix blob.ID, callba
 	return nil
 }
 
-// Close is a no-op for this implementation.
-func (s *objectLockingMap) Close(ctx context.Context) error {
-	return nil
-}
-
 // TouchBlob updates the mtime of the latest version of the object. If it is a
 // delete-marker or if it does not exist then this becomes a no-op. If the
 // latest version has retention parameters set then they are respected.
@@ -286,14 +297,9 @@ func (s *objectLockingMap) DisplayName() string {
 	return "VersionedMap"
 }
 
-// FlushCaches is a no-op for this implementation.
-func (s *objectLockingMap) FlushCaches(ctx context.Context) error {
-	return nil
-}
-
 // NewVersionedMapStorage returns an implementation of Storage backed by the
 // contents of an internal in-memory map used primarily for testing.
-func NewVersionedMapStorage(timeNow func() time.Time) blob.Storage {
+func NewVersionedMapStorage(timeNow func() time.Time) RetentionStorage {
 	if timeNow == nil {
 		timeNow = clock.Now
 	}
