@@ -62,10 +62,8 @@ func (c *contentCacheImpl) GetContent(ctx context.Context, contentID string, blo
 }
 
 func (c *contentCacheImpl) getContentFromFullBlob(ctx context.Context, blobID blob.ID, offset, length int64, output *gather.WriteBuffer) error {
-	// acquire exclusive lock
-	mut := c.pc.GetFetchingMutex(blobID)
-	mut.Lock()
-	defer mut.Unlock()
+	c.pc.exclusiveLock(string(blobID))
+	defer c.pc.exclusiveUnlock(string(blobID))
 
 	// check again to see if we perhaps lost the race and the data is now in cache.
 	if c.pc.GetPartial(ctx, BlobIDCacheKey(blobID), offset, length, output) {
@@ -114,9 +112,8 @@ func (c *contentCacheImpl) fetchBlobInternal(ctx context.Context, blobID blob.ID
 
 func (c *contentCacheImpl) getContentFromFullOrPartialBlob(ctx context.Context, contentID string, blobID blob.ID, offset, length int64, output *gather.WriteBuffer) error {
 	// acquire shared lock on a blob, PrefetchBlob will acquire exclusive lock here.
-	mut := c.pc.GetFetchingMutex(blobID)
-	mut.RLock()
-	defer mut.RUnlock()
+	c.pc.sharedLock(string(blobID))
+	defer c.pc.sharedUnlock(string(blobID))
 
 	// see if we have the full blob cached by extracting a partial range.
 	if c.pc.GetPartial(ctx, BlobIDCacheKey(blobID), offset, length, output) {
@@ -124,9 +121,8 @@ func (c *contentCacheImpl) getContentFromFullOrPartialBlob(ctx context.Context, 
 	}
 
 	// acquire exclusive lock on the content
-	mut2 := c.pc.GetFetchingMutex(blob.ID(contentID))
-	mut2.Lock()
-	defer mut2.Unlock()
+	c.pc.exclusiveLock(contentID)
+	defer c.pc.exclusiveUnlock(contentID)
 
 	output.Reset()
 
@@ -161,9 +157,8 @@ func (c *contentCacheImpl) PrefetchBlob(ctx context.Context, blobID blob.ID) err
 	}
 
 	// acquire exclusive lock for the blob.
-	mut := c.pc.GetFetchingMutex(blobID)
-	mut.Lock()
-	defer mut.Unlock()
+	c.pc.exclusiveLock(string(blobID))
+	defer c.pc.exclusiveUnlock(string(blobID))
 
 	if c.pc.GetPartial(ctx, BlobIDCacheKey(blobID), 0, 1, &blobData) {
 		return nil
