@@ -12,12 +12,19 @@ type Options struct {
 	// Prefix specifies additional string to prepend to all objects.
 	Prefix string `json:"prefix,omitempty"`
 
-	// Azure Storage account name and key
+	// Storage account name
 	StorageAccount string `json:"storageAccount"`
-	StorageKey     string `json:"storageKey"     kopia:"sensitive"`
+
+	// Storage account access key
+	StorageKey string `json:"storageKey"     kopia:"sensitive"`
 
 	// Alternatively provide SAS Token
 	SASToken string `json:"sasToken" kopia:"sensitive"`
+
+	// the tenant-ID/client-ID/client-Secret of the service principal
+	TenantID     string
+	ClientID     string
+	ClientSecret string
 
 	StorageDomain string `json:"storageDomain,omitempty"`
 
