@@ -34,7 +34,7 @@ const (
 
 type fsStorage struct {
 	sharded.Storage
-	blob.UnsupportedBlobRetention
+	blob.DefaultProviderImplementation
 }
 
 type fsImpl struct {
@@ -316,8 +316,6 @@ func (fs *fsStorage) TouchBlob(ctx context.Context, blobID blob.ID, threshold ti
 
 		mtime = n
 
-		log(ctx).Debugf("updating timestamp on %v to %v", path, n)
-
 		//nolint:wrapcheck
 		return osi.Chtimes(path, n, n)
 	}, fs.Impl.(*fsImpl).isRetriable)
@@ -334,14 +332,6 @@ func (fs *fsStorage) ConnectionInfo() blob.ConnectionInfo {
 
 func (fs *fsStorage) DisplayName() string {
 	return fmt.Sprintf("Filesystem: %v", fs.RootPath)
-}
-
-func (fs *fsStorage) Close(ctx context.Context) error {
-	return nil
-}
-
-func (fs *fsStorage) FlushCaches(ctx context.Context) error {
-	return nil
 }
 
 // New creates new filesystem-backed storage in a specified directory.
