@@ -138,12 +138,37 @@ type ExtendOptions struct {
 	RetentionPeriod time.Duration
 }
 
-// UnsupportedBlobRetention provides a default implementation for ExtendBlobRetention.
-type UnsupportedBlobRetention struct{}
+// DefaultProviderImplementation provides a default implementation for
+// common functions that are mostly provider independent and have a sensible
+// default.
+//
+// Storage providers should imbed this struct and override functions that they
+// have different return values for.
+type DefaultProviderImplementation struct{}
 
 // ExtendBlobRetention provides a common implementation for unsupported blob retention storage.
-func (s *UnsupportedBlobRetention) ExtendBlobRetention(context.Context, ID, ExtendOptions) error {
+func (s DefaultProviderImplementation) ExtendBlobRetention(context.Context, ID, ExtendOptions) error {
 	return ErrUnsupportedObjectLock
+}
+
+// IsReadOnly complies with the Storage interface.
+func (s DefaultProviderImplementation) IsReadOnly() bool {
+	return false
+}
+
+// Close complies with the Storage interface.
+func (s DefaultProviderImplementation) Close(context.Context) error {
+	return nil
+}
+
+// FlushCaches complies with the Storage interface.
+func (s DefaultProviderImplementation) FlushCaches(context.Context) error {
+	return nil
+}
+
+// GetCapacity complies with the Storage interface.
+func (s DefaultProviderImplementation) GetCapacity(context.Context) (Capacity, error) {
+	return Capacity{}, ErrNotAVolume
 }
 
 // HasRetentionOptions returns true when blob-retention settings have been
@@ -182,6 +207,10 @@ type Storage interface {
 
 	// ExtendBlobRetention extends the retention time of a blob (when blob retention is enabled)
 	ExtendBlobRetention(ctx context.Context, blobID ID, opts ExtendOptions) error
+
+	// IsReadOnly returns whether this Storage is in read-only mode. When in
+	// read-only mode all mutation operations will fail.
+	IsReadOnly() bool
 }
 
 // ID is a string that represents blob identifier.
