@@ -9,9 +9,21 @@ import (
 func Broadcast(logger ...Logger) Logger {
 	var cores []zapcore.Core
 
+	var singleName string
+
 	for _, l := range logger {
-		cores = append(cores, l.Desugar().Core())
+		dl := l.Desugar()
+
+		if singleName == "" {
+			singleName = dl.Name()
+		}
+
+		if dl.Name() != singleName {
+			singleName = "-"
+		}
+
+		cores = append(cores, dl.Core())
 	}
 
-	return zap.New(zapcore.NewTee(cores...)).Sugar()
+	return zap.New(zapcore.NewTee(cores...)).Sugar().Named(singleName)
 }
