@@ -34,6 +34,15 @@ func WithLogger(ctx context.Context, l LoggerFactory) context.Context {
 	})
 }
 
+// WithAdditionalLogger returns a context where all logging is emitted the original output plus the provided logger factory.
+func WithAdditionalLogger(ctx context.Context, fact LoggerFactory) context.Context {
+	originalLogFactory := loggerFactoryFromContext(ctx)
+
+	return WithLogger(ctx, func(module string) Logger {
+		return Broadcast(originalLogFactory(module), fact(module))
+	})
+}
+
 // loggerFactoryFromContext returns a LoggerFactory associated with current context.
 func loggerFactoryFromContext(ctx context.Context) LoggerFactory {
 	v := ctx.Value(loggerCacheKey)
@@ -42,13 +51,4 @@ func loggerFactoryFromContext(ctx context.Context) LoggerFactory {
 	}
 
 	return v.(*loggerCache).getLogger //nolint:forcetypeassert
-}
-
-// AlsoLogTo returns a context where all logging is emitted the original output plus the provided loggers.
-func AlsoLogTo(ctx context.Context, loggers ...Logger) context.Context {
-	originalLogFactory := loggerFactoryFromContext(ctx)
-
-	return WithLogger(ctx, func(module string) Logger {
-		return Broadcast(append([]Logger{originalLogFactory(module)}, loggers...)...)
-	})
 }
