@@ -10,11 +10,13 @@ import (
 	"time"
 
 	"github.com/pkg/errors"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/kopia/kopia/internal/blobtesting"
 	"github.com/kopia/kopia/internal/testlogging"
 	"github.com/kopia/kopia/internal/testutil"
+	"github.com/kopia/kopia/repo/blob/readonly"
 	"github.com/kopia/kopia/repo/content"
 	"github.com/kopia/kopia/repo/content/index"
 	"github.com/kopia/kopia/repo/encryption"
@@ -27,7 +29,7 @@ func TestMain(m *testing.M) { testutil.MyTestMain(m) }
 func TestManifest(t *testing.T) {
 	ctx := testlogging.Context(t)
 	data := blobtesting.DataMap{}
-	mgr := newManagerForTesting(ctx, t, data)
+	mgr := newManagerForTesting(ctx, t, data, ManagerOptions{})
 
 	item1 := map[string]int{"foo": 1, "bar": 2}
 	item2 := map[string]int{"foo": 2, "bar": 3}
@@ -87,7 +89,7 @@ func TestManifest(t *testing.T) {
 
 	// flush underlying content manager and verify in new manifest manager.
 	mgr.b.Flush(ctx)
-	mgr2 := newManagerForTesting(ctx, t, data)
+	mgr2 := newManagerForTesting(ctx, t, data, ManagerOptions{})
 
 	for _, tc := range cases {
 		verifyMatches(ctx, t, mgr2, tc.criteria, tc.expected)
@@ -138,7 +140,7 @@ func TestManifest(t *testing.T) {
 
 	mgr.b.Flush(ctx)
 
-	mgr3 := newManagerForTesting(ctx, t, data)
+	mgr3 := newManagerForTesting(ctx, t, data, ManagerOptions{})
 
 	verifyItem(ctx, t, mgr3, id1, labels1, item1)
 	verifyItem(ctx, t, mgr3, id2, labels2, item2)
@@ -306,10 +308,18 @@ func sortIDs(s []ID) {
 	})
 }
 
-func newManagerForTesting(ctx context.Context, t *testing.T, data blobtesting.DataMap) *Manager {
+type contentManagerOpts struct {
+	readOnly bool
+}
+
+func newContentManagerForTesting(ctx context.Context, t *testing.T, data blobtesting.DataMap, opts contentManagerOpts) contentManager {
 	t.Helper()
 
 	st := blobtesting.NewMapStorage(data, nil, nil)
+
+	if opts.readOnly {
+		st = readonly.NewWrapper(st)
+	}
 
 	fop, err := format.NewFormattingOptionsProvider(&format.ContentFormat{
 		Hash:       hashing.DefaultAlgorithm,
@@ -327,7 +337,15 @@ func newManagerForTesting(ctx context.Context, t *testing.T, data blobtesting.Da
 
 	t.Cleanup(func() { bm.CloseShared(ctx) })
 
-	mm, err := NewManager(ctx, bm, ManagerOptions{}, nil)
+	return bm
+}
+
+func newManagerForTesting(ctx context.Context, t *testing.T, data blobtesting.DataMap, options ManagerOptions) *Manager {
+	t.Helper()
+
+	bm := newContentManagerForTesting(ctx, t, data, contentManagerOpts{})
+
+	mm, err := NewManager(ctx, bm, options, nil)
 	require.NoError(t, err)
 
 	return mm
@@ -336,7 +354,7 @@ func newManagerForTesting(ctx context.Context, t *testing.T, data blobtesting.Da
 func TestManifestInvalidPut(t *testing.T) {
 	ctx := testlogging.Context(t)
 	data := blobtesting.DataMap{}
-	mgr := newManagerForTesting(ctx, t, data)
+	mgr := newManagerForTesting(ctx, t, data, ManagerOptions{})
 
 	cases := []struct {
 		labels        map[string]string
@@ -359,7 +377,7 @@ func TestManifestAutoCompaction(t *testing.T) {
 	ctx := testlogging.Context(t)
 	data := blobtesting.DataMap{}
 
-	mgr := newManagerForTesting(ctx, t, data)
+	mgr := newManagerForTesting(ctx, t, data, ManagerOptions{})
 
 	for i := 0; i < 100; i++ {
 		item1 := map[string]int{"foo": 1, "bar": 2}
@@ -380,4 +398,95 @@ func TestManifestAutoCompaction(t *testing.T) {
 		require.NoError(t, mgr.Flush(ctx))
 		require.NoError(t, mgr.b.Flush(ctx))
 	}
+}
+
+func TestManifestConfigureAutoCompaction(t *testing.T) {
+	ctx := testlogging.Context(t)
+	data := blobtesting.DataMap{}
+	item1 := map[string]int{"foo": 1, "bar": 2}
+	labels1 := map[string]string{"type": "item", "color": "red"}
+	compactionCount := 99
+
+	mgr := newManagerForTesting(ctx, t, data, ManagerOptions{AutoCompactionThreshold: compactionCount})
+
+	for i := 0; i < compactionCount-1; i++ {
+		addAndVerify(ctx, t, mgr, labels1, item1)
+		require.NoError(t, mgr.Flush(ctx))
+		require.NoError(t, mgr.b.Flush(ctx))
+	}
+
+	// Should not trigger compaction
+	_, err := mgr.Find(ctx, labels1)
+	require.NoError(t, err)
+
+	foundContents := getManifestContentCount(ctx, t, mgr)
+
+	if got, want := foundContents, compactionCount-1; got != want {
+		t.Errorf("unexpected number of blocks: %v, want %v", got, want)
+	}
+
+	// Add another manifest
+	addAndVerify(ctx, t, mgr, labels1, item1)
+
+	require.NoError(t, mgr.Flush(ctx))
+	require.NoError(t, mgr.b.Flush(ctx))
+
+	// *Should* trigger compaction
+	_, err = mgr.Find(ctx, labels1)
+	require.NoError(t, err)
+
+	foundContents = getManifestContentCount(ctx, t, mgr)
+
+	if got, want := foundContents, 1; got != want {
+		t.Errorf("unexpected number of blocks: %v, want %v", got, want)
+	}
+}
+
+func getManifestContentCount(ctx context.Context, t *testing.T, mgr *Manager) int {
+	t.Helper()
+
+	foundContents := 0
+
+	if err := mgr.b.IterateContents(
+		ctx,
+		content.IterateOptions{Range: index.PrefixRange(ContentPrefix)},
+		func(ci content.Info) error {
+			foundContents++
+			return nil
+		}); err != nil {
+		t.Errorf("unable to list manifest content: %v", err)
+	}
+
+	return foundContents
+}
+
+func TestManifestAutoCompactionWithReadOnly(t *testing.T) {
+	ctx := testlogging.Context(t)
+	data := blobtesting.DataMap{}
+
+	bm := newContentManagerForTesting(ctx, t, data, contentManagerOpts{})
+
+	mgr, err := NewManager(ctx, bm, ManagerOptions{}, nil)
+	require.NoError(t, err, "getting initial manifest manager")
+
+	for i := 0; i < 100; i++ {
+		item1 := map[string]int{"foo": 1, "bar": 2}
+		labels1 := map[string]string{"type": "item", "color": "red"}
+
+		_, err = mgr.Put(ctx, labels1, item1)
+		require.NoError(t, err, "adding item to manifest manager")
+
+		require.NoError(t, mgr.Flush(ctx))
+		require.NoError(t, mgr.b.Flush(ctx))
+	}
+
+	// Opening another instance of the manager should cause the manifest manager
+	// to attempt to compact things.
+	bm = newContentManagerForTesting(ctx, t, data, contentManagerOpts{readOnly: true})
+
+	mgr, err = NewManager(ctx, bm, ManagerOptions{}, nil)
+	require.NoError(t, err, "getting other instance of manifest manager")
+
+	_, err = mgr.Find(ctx, map[string]string{"color": "red"})
+	assert.NoError(t, err, "forcing reload of manifest manager")
 }
